python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-proxy"
version = "0.1.0"
description = "LiteLLM proxy with per-agent model routing and OAuth token injection"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "fullstacktard"}
]
keywords = ["litellm", "claude", "proxy", "agent-routing", "oauth"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "litellm[proxy]>=1.50.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
]

[project.scripts]
claude-proxy = "claude_proxy.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["claude_proxy"]

[tool.hatch.build.targets.sdist]
include = [
    "/claude_proxy",
    "requirements.txt"
]
